ikev2: add option to specify multiple destinations for redirection during IKE_SA_INIT
This addition enables to have one 'front' server that will be used as a load
balancer, redirecting clients to a list of other servers.

Before it was possible only to have something like 'global-redirect-to=1.1.1.1'
Now one can have, i.e., 'global-redirect-to=1.1.1.1,2.2.2.2,3.3.3.3'.

functions added are supposed to be modular, so that in the future we can easily
add the possibility of specifying multiple destinations for redirection
during IKE_AUTH as well.