ikev2: redirect: refactor redirect during active sessions
Joint work with Paul Wouters

- remove --peer-ip <ip_address> from whack --redirect command, whose itended
use case was to redirect a *single* peer to other gateway. That is basically
useless, so we remove it.
- improve logging
- update ikev2-redirect-06-roadwarriors test. Note: False negatives can happen
in this test, because the roadwarriors are redirected at the same time
and sometimes they arive in the right order to the other gateway, sometimes
they arrive in 'wrong' order which causes inverted console logs and test fail.